+++ FUSOU-PROXY/proxy_rust/proxy/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod pac_server;
-pub mod proxy_server;
+pub mod proxy_server_http;
 pub mod cmd_pac;
 pub mod edit_pac;
 pub mod bidirectional_channel;
